// backend/src/types/graphql.ts
export type Maybe<T> = T | undefined;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  Datetime: any;
  UUID: string;
  Url: any;
  _FieldSet: any;
};
// ...
export type ProductCategory = {
  __typename?: 'ProductCategory';
  description: Scalars['String'];
  id: Scalars['UUID'];
  name: Scalars['String'];
};
// ...
export type UserInventory = {
  __typename?: 'UserInventory';
  id: Scalars['UUID'];
  items: Array<Maybe<UserInventoryItem>>;
};

export type UserInventoryItem = {
  __typename?: 'UserInventoryItem';
  category?: Maybe<ProductCategory>;
  consumedAt?: Maybe<Scalars['Date']>;
  description: Scalars['String'];
  id: Scalars['UUID'];
  imageUrl: Scalars['String'];
  name: Scalars['String'];
  paidAt: Scalars['Date'];
  paidPrice: Scalars['Float'];
  variant?: Maybe<Scalars['String']>;
};